refactor: move browse filter components
