+++ src/app/(core)/teachers/programmes/[slug]/[tab]/Components/Filters/BrowseFilters/BrowseFiltersSubjectCategories.fixtures.ts
