+++ src/app/(core)/teachers/programmes/[slug]/[tab]/Components/Filters/BrowseFilters/BrowseFiltersYears.test.tsx
@@ -1,7 +1,8 @@
 import { act } from "@testing-library/react";
 
+import { basicSetup } from "../../../../../../../../../components/CurriculumComponents/CurricVisualiserFilters/CurricFiltersYears.fixtures";
+
 import { BrowseFiltersYears } from "./BrowseFiltersYears";
-import { basicSetup } from "./CurricFiltersYears.fixtures";
 
 import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
 import { BrowseFiltersProvider } from "@/context/BrowseFilters";
