+++ src/components/CurriculumComponents/CurricVisualiserFilters/index.tsx
@@ -1,6 +1,6 @@
 export { CurricFiltersThreads } from "./CurricFiltersThreads";
 export { CurricFiltersYears } from "./CurricFiltersYears";
-export { BrowseFiltersYears } from "./BrowseFiltersYears";
-export { BrowseFiltersTiers } from "./BrowseFiltersTiers";
-export { BrowseFiltersSubjectCategories } from "./BrowseFiltersSubjectCategories";
-export { BrowseFiltersChildSubjects } from "./BrowseFiltersChildSubjects";
+export { BrowseFiltersYears } from "../../../app/(core)/teachers/programmes/[slug]/[tab]/Components/Filters/BrowseFilters/BrowseFiltersYears";
+export { BrowseFiltersTiers } from "../../../app/(core)/teachers/programmes/[slug]/[tab]/Components/Filters/BrowseFilters/BrowseFiltersTiers";
+export { BrowseFiltersSubjectCategories } from "../../../app/(core)/teachers/programmes/[slug]/[tab]/Components/Filters/BrowseFilters/BrowseFiltersSubjectCategories";
+export { BrowseFiltersChildSubjects } from "../../../app/(core)/teachers/programmes/[slug]/[tab]/Components/Filters/BrowseFilters/BrowseFiltersChildSubjects";
